% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demography_functions.R
\name{create_data_input}
\alias{create_data_input}
\title{Create empty template for demography life history data}
\usage{
create_data_input(maturity.type, t0 = FALSE)
}
\arguments{
\item{maturity.type}{The type of maturity estimates from the source life history study.
must be one of `logistic - int/slope`, `logistic - a50/a95`, `normal` or `uniform`.}

\item{t0}{Logical argument regarding whether the growth models included "t0" or "L0" as
as a parameter. Default is `FALSE`}
}
\value{
A multi-level list of the class `Demography.inputs`
}
\description{
The demography functions in this package require a large amount of detailed
    life history information. This is provided to these functions as a multi-level list
    of various life history parameters with the class `Demography.inputs`. This function
    creates the template for this input data which can be filled in after it is created.
}
