% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharkDemography.R
\docType{package}
\name{SharkDemography.}
\alias{SharkDemography.}
\title{SharkDemography.}
\description{
A package to conduct Shark demographic analyses using Leslie matrix models.
}
\author{
Jonathan Smart
}
