% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharkDemography.R
\docType{data}
\name{Silky_data}
\alias{Silky_data}
\title{Example life history data for Silky Sharks}
\format{
A list of 10 with class "Demography.inputs"
}
\usage{
data(Silky_data)
}
\description{
An example dataset of a completed set of life history data for Silky Sharks
in a `create_data_input()` template.
}
\keyword{datasets}
